% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custfunctions.R
\name{ETC}
\alias{ETC}
\title{ETC: Function to run equivalence testing on single correlation = 1}
\usage{
ETC(dat, varA, varB, ei, alpha = 0.05)
}
\arguments{
\item{dat}{Matrix object representing the data, subjects as rows, variables as columns.}

\item{varA}{Column number corresponding to one variable of the pair being tested for redundancy.}

\item{varB}{Column number corresponding to the other variable of the pair being tested for redundancy.}

\item{ei}{Single value indicating the width of the equivalence interval, where 1-\code{"ei"} represents the lower bound of this interval and 1 represents the upper bound.}

\item{alpha}{Single value representing the desired alpha level.}
}
\value{
A list that includes a z-statistic, p-value, and redundancy decision, in addition to select input parameters for convenience.
}
\description{
ETC: Function to run equivalence testing on single correlation = 1
}
\details{
For a single pair of variables, this function runs equivalence testing on a single zero-order correlation. Specifically, the test evaluates whether the variables' correlation is close enough to 1 as to declare the variables effectively redundant. The inspiration for this function came from work by Goertzen & Cribbie (2010).
}
\examples{
\dontrun{
#### 1. Example with bfi dataset from psych package

   ## Clean dataset to get complete data only (n = 2436, and 25 items)
   # Needs psych package loaded in
   library("psych")
   data(bfi)
   bfisub = bfi[,1:25]
   # Remove any rows with missing data
   bfisub <- na.omit(bfisub)

   ## Pick alpha level and equivalence interval
   alpha = .05
   ei = .1

   ## Run ETC test on single pair of items (N1 and N2)
   ETC <- ETC(dat=bfisub, varA=16, varB=17, alpha=alpha, ei=ei)
}

}
\references{
Goertzen, J. R., & Cribbie, R. A. (2010). Detecting a lack of association: An equivalence testing approach. British Journal of Mathematical and Statistical Psychology, 63(3), 527-537.

Some older refs too
}
