% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custfunctions.R
\name{ETT}
\alias{ETT}
\title{ETT: Function to run equivalence testing on difference between two dependent correlations}
\usage{
ETT(dat, varA, varB, ei, alpha = 0.05, threshold = 0.75, corMin = 0.5)
}
\arguments{
\item{dat}{Matrix object representing the data, subjects as rows, variables as columns.}

\item{varA}{Column number corresponding to one variable of the pair being tested for redundancy.}

\item{varB}{Column number corresponding to the other variable of the pair being tested for redundancy.}

\item{ei}{Single value indicating half the width of the equivalence interval, where -\code{"ei"} represents the lower bound of this interval and \code{"ei"} represents the upper bound (i.e., 0 is the midpoint).}

\item{alpha}{Single value representing the desired alpha level.}

\item{threshold}{Single value between 0 and 1, representing the minimum proportion of rejections required for the two variables to be declared redundant.}

\item{corMin}{Single value between 0 and 1, representing the minimum zero-order correlation required (between \code{"varA"} and \code{"varB"}) for the two variables to be declared redundant.}
}
\value{
A list that includes results for all tests including p-values, as well as the proportion of tests resulting in rejection of the null hypothesis, and the redundancy decision, in addition to select input parameters for convenience.
}
\description{
ETT: Function to run equivalence testing on difference between two dependent correlations
}
\details{
For a single pair of variables, this function runs equivalence testing on the topology of the variables - in other words, the difference between the correlations of \code{"varA"} and \code{"varB"} with each of the other variables in the dataset. Specifically, the test evaluates whether this difference is within an user-defined interval around zero such that the variables can be declared redundant. Note that multiple tests are conducted for a given variable pair when the number of variables in the dataset exceeds three. The inspiration for this function came from work by Counsell & Cribbie (2015).
}
\examples{
\dontrun{
#### 1. Example with bfi dataset from psych package

   ## Clean dataset to get complete data only (n = 2436, and 25 items)
   # Needs psych package loaded in
   library("psych")
   data(bfi)
   bfisub = bfi[,1:25]
   # Remove any rows with missing data
   bfisub <- na.omit(bfisub)

   ## Pick alpha level, equivalence interval, proportion rejected threshold (ETT),
   ## and minimum zero-order correlation (ETT)
   alpha = .05
   ei = .1
   threshold = .75
   corMin = .5

   ## Run ETC test on single pair of items (N1 and N2)
   ETT <- ETT(dat=bfisub, varA=16, varB=17, alpha=alpha, ei=ei,
     threshold=threshold, corMin=corMin)
}

}
\references{
Counsell, A., & Cribbie, R. A. (2015). Equivalence tests for comparing correlation and regression coefficients. British Journal of Mathematical and Statistical Psychology, 68(2), 292-309.

Some older refs too
}
