% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relfuns.R
\name{fit_redmod}
\alias{fit_redmod}
\title{Fits two alternative (b)lavaan models for purposes of comparing item overlap}
\usage{
fit_redmod(
  syn,
  dat,
  software = c("lavaan", "blavaan"),
  likelihood = "wishart",
  std.lv = TRUE,
  auto.fix.first = FALSE,
  ...
)
}
\arguments{
\item{syn}{syntax of two models, as created by \code{syn_redmod}}

\item{dat}{dataset; should be a data frame with column labels. Should also
only contain the variables of interest.}

\item{software}{which software package to use?}

\item{likelihood}{for \code{lavaan}, which likelihood to use? (Used to
override default behavior)}

\item{std.lv}{for \code{lavaan} and \code{blavaan}, whether to standardize
latent variables (Used to override default behavior)}

\item{auto.fix.first}{for \code{lavaan} and \code{blabvaan}, whether to fix
loading for first item of each latent variable to one (Used to override
default behavior)}

\item{...}{Additional arguments passed to \code{lavaan} or \code{blavaan}}
}
\description{
Fits two alternative (b)lavaan models for purposes of comparing item overlap
}
