% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relfuns.R
\name{syn_redmod}
\alias{syn_redmod}
\title{Create (b)lavaan syntax to fit two alternative models}
\usage{
syn_redmod(
  dat,
  v1,
  v2,
  method = c("lv", "pconst", "sat"),
  warmstart = FALSE,
  corconst = NULL
)
}
\arguments{
\item{dat}{dataset; should be a data frame with column labels. Should also
only contain the variables of interest.}

\item{v1}{name of variable 1}

\item{v2}{name of variable 2}

\item{method}{which type of model should we estimate?
lv = latent variable model to test topology. pconst = implement
  proportionality constraints on topology.
sat = just does saturated model (e.g., in case one wants only to check for
  some constraint on the correlation b/w variables)}

\item{warmstart}{use sample means and covariances to help w/ starting values?}

\item{corconst}{if there should be any constraint on the model-implied
correlation between the two variables, put the value here. Defaults to no
such constraint (i.e., test only topology)}
}
\description{
One model is saturated; other assumes two items measure same construct
}
\examples{
\donttest{
# Example with bfi dataset from psych package
library(psych)
bfisub <- bfi[,1:25]

# remove any rows with missing data
bfisub <- na.omit(bfisub)

syntax <- syn_redmod(bfisub, "A1", "A2")
cat(syntax$mod) # syntax for restricted model
cat(syntax$modsat) # syntax for saturated model
}
}
