% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relfuns.R
\name{sel_redmod}
\alias{sel_redmod}
\title{Does model selection}
\usage{
sel_redmod(fits, index = c("aic", "bic", "sic", "waic", "loo"), thresh = 0)
}
\arguments{
\item{fits}{two fitted models, as created by \code{syn_redmod} then
\code{fit_redmod}}

\item{index}{which fit index to examine? Options depend on whether
\code{lavaan} or \code{blavaan} was used to estimate the models.}

\item{thresh}{threshold for determining whether the saturated model wins.
Positive values result in a stronger preference for parsimony (i.e., items
are redundant)}
}
\description{
Does model selection
}
